model_workers/safety_guard/requirements.txt
ckip_transformers==0.3.4
accelerate
torch
transformers
protobuf
pyahocorasick
langchain
langchain-community
sentence-transformers
faiss-cpu